requirements.txt
requests>=2.31.0
aiogram~=3.17.0
pydantic~=2.10.5
telebot~=0.0.5
colorama~=0.4.6
openai~=1.59.8
pyTelegramBotAPI>=4.12.0
python-dotenv>=1.0.0
supabase>=1.0.3
python-dateutil>=2.8.2